chore(mobile): trim subscription widget scaffolding
