+++ apps/mobile/package.json
@@ -42,7 +42,6 @@
     "sync:pierre-icons": "node modules/t3-markdown-text/scripts/sync-pierre-file-icons.mjs",
     "generate": "node --disable-warning=MODULE_TYPELESS_PACKAGE_JSON scripts/generate-uniwind-themes.mts",
     "test": "vp test run",
-    "test:widget-runtime": "node scripts/verify-subscription-widget.mjs",
     "typecheck": "tsc --noEmit"
   },
   "dependencies": {
@@ -132,9 +131,7 @@
     "@pierre/trees": "1.0.0-beta.4",
     "@types/react": "~19.2.0",
     "@types/react-dom": "~19.2.3",
-    "@types/react-test-renderer": "19.1.0",
     "babel-preset-expo": "~57.0.9",
-    "react-test-renderer": "19.2.3",
     "tailwindcss": "^4.0.0",
     "typescript": "catalog:"
   },
